Site Review and Tweaks

+++ app/views/users/partials/follow-form.blade.php
@@ -1,19 +1,21 @@
-@if($user->isFollowedBy($currentUser))
+@if($signedIn)
+	@if($user->isFollowedBy($currentUser))
 
-	{{ Form::open(['method' => 'DELETE', 'route' => ['follow_path', $user->id]]) }}
+		{{ Form::open(['method' => 'DELETE', 'route' => ['follow_path', $user->id]]) }}
 
-		{{ Form::hidden('userIdToUnfollow', $user->id) }}
-		<button class="btn btn-danger" type="submit">Unfollow {{ $user->username }}</button>
+			{{ Form::hidden('userIdToUnfollow', $user->id) }}
+			<button class="btn btn-danger" type="submit">Unfollow {{ $user->username }}</button>
 
-	{{ Form::close() }}
+		{{ Form::close() }}
 
-@else
+	@else
 
-	{{ Form::open(['route' => 'follows_path']) }}
+		{{ Form::open(['route' => 'follows_path']) }}
 
-		{{ Form::hidden('userIdToFollow', $user->id) }}
-		<button class="btn btn-primary" type="submit">Follow {{ $user->username }}</button>
+			{{ Form::hidden('userIdToFollow', $user->id) }}
+			<button class="btn btn-primary" type="submit">Follow {{ $user->username }}</button>
 
-	{{ Form::close() }}
+		{{ Form::close() }}
 
+	@endif
 @endif
